\name{setup_dir}
\alias{setup_dir}
\title{
Setup working directory
}
\description{
Choose a directory from the pop-up window to save the final output (csv files) created by \code{run()}.
}
\usage{
setup_dir()
}
\details{
The function is a wrapper for \code{setwd}.
}
\value{
Returns the location of the directory chosen.
}
\seealso{
\code{\link{setwd}}, \code{\link{choose_dir}}
}
\examples{
## NOT RUN
# setup_dir()
}
