\name{run}
\alias{run}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compare other packages' functions for nonlinear least squares with best reference values
}
\description{
Perform comparison of best reference values of parameters and sum of squares in a nonlinear problem with other packages' functions for nonlinear least squares.

}
\usage{
run(machid, database, errorlog)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{machid}{
Takes the machine ID generated by \code{machineId()}.
}
  \item{database}{
Takes the dataframe generated by \code{create_db()}.
}
  \item{errorlog}{
Takes the dataframe generated by \code{create_elog()}.
}
}
\value{
It returns a list of 2 items:
\item{database}{updated database dataframe}
\item{errorlog}{updated error log dataframe}
}

